+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 - Add Dashboard router
 - Add generic dashboard component
 - Add dashboard components for Order, Enrollment
+- Add dashboard components for Credit Card management
 
 ### Changed